Rename "snapfs" to "squashfs"
Gustavo suggested this rename in an earlier code review and it
does make sense, its a squashfs based snap.
