+++ pkg/squashfs/squashfs.go
@@ -17,7 +17,7 @@
  *
  */
 
-package snapfs
+package squashfs
 
 import (
 	"fmt"
@@ -40,12 +40,12 @@ func (s *Snap) Name() string {
 	return filepath.Base(s.path)
 }
 
-// New returns a new Snapfs snap
+// New returns a new Squashfs snap
 func New(path string) *Snap {
 	return &Snap{path: path}
 }
 
-// Close is not doing anything for snapfs - COMPAT
+// Close is not doing anything for squashfs - COMPAT
 func (s *Snap) Close() error {
 	return nil
 }
@@ -60,7 +60,7 @@ func (s *Snap) MetaMember(metaMember string) ([]byte, error) {
 	return s.ReadFile(filepath.Join("meta", metaMember))
 }
 
-// ExtractHashes does notthing for snapfs snaps - COMAPT
+// ExtractHashes does notthing for squashfs snaps - COMAPT
 func (s *Snap) ExtractHashes(dir string) error {
 	return nil
 }
@@ -94,7 +94,7 @@ func (s *Snap) Unpack(src, dstDir string) error {
 	return runCommand("unsquashfs", "-f", "-i", "-d", dstDir, s.path, src)
 }
 
-// ReadFile returns the content of a single file inside a snapfs snap
+// ReadFile returns the content of a single file inside a squashfs snap
 func (s *Snap) ReadFile(path string) (content []byte, err error) {
 	tmpdir, err := ioutil.TempDir("", "read-file")
 	if err != nil {
@@ -118,7 +118,7 @@ func (s *Snap) CopyBlob(targetFile string) error {
 
 // Verify verifies the snap
 func (s *Snap) Verify(unauthOk bool) error {
-	// FIXME: there is no verification yet for snapfs packages, this
+	// FIXME: there is no verification yet for squashfs packages, this
 	//        will be done via assertions later for now we rely on
 	//        the https security
 	return nil
