+++ snappy/build_test.go
@@ -500,7 +500,7 @@ version: 1.0.1
 	c.Assert(err, ErrorMatches, "can not handle type of file .*")
 }
 
-func (s *SnapTestSuite) TestBuildSnapfsSimple(c *C) {
+func (s *SnapTestSuite) TestBuildSquashfsSimple(c *C) {
 	sourceDir := makeExampleSnapSourceDir(c, `name: hello
 version: 1.0.1
 architecture: ["i386", "amd64"]
@@ -509,7 +509,7 @@ integration:
   apparmor-profile: meta/hello.apparmor
 `)
 
-	resultSnap, err := BuildSnapfsSnap(sourceDir, "")
+	resultSnap, err := BuildSquashfsSnap(sourceDir, "")
 	c.Assert(err, IsNil)
 	defer os.Remove(resultSnap)
